# src/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: |
    Order Microservice – manages shopping basket operations, checkout,
    and order history. Publishes `ORDER_COMPLETED` events to Kafka
    on successful checkout.
  version: 1.0.0
servers:
  - url: http://localhost:3002
    description: Direct service access
  - url: http://localhost:3000
    description: Via API Gateway
tags:
  - name: Health
    description: Service health check
  - name: Basket
    description: Shopping basket operations
  - name: Orders
    description: Order management and checkout

components:
  schemas:
    BasketItem:
      type: object
      properties:
        cakeId:
          type: string
          example: 667a1b2c3d4e5f6a7b8c9d0e
        name:
          type: string
          example: Velvet Midnight
        price:
          type: number
          example: 32
        quantity:
          type: integer
          example: 2

    Basket:
      type: object
      properties:
        _id:
          type: string
        customerId:
          type: string
          example: demo-customer
        items:
          type: array
          items:
            $ref: "#/components/schemas/BasketItem"
        total:
          type: number
          example: 64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AddToBasketRequest:
      type: object
      required: [cakeId]
      properties:
        cakeId:
          type: string
          example: 667a1b2c3d4e5f6a7b8c9d0e
        quantity:
          type: integer
          minimum: 1
          default: 1
          example: 2

    UpdateQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          example: 3

    CheckoutRequest:
      type: object
      required: [customerEmail]
      properties:
        customerEmail:
          type: string
          format: email
          example: user@cakedelight.com

    Order:
      type: object
      properties:
        _id:
          type: string
        customerId:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/BasketItem"
        total:
          type: number
        status:
          type: string
          enum: [CONFIRMED, PROCESSING, DELIVERED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

  parameters:
    XCustomerId:
      name: x-customer-id
      in: header
      schema:
        type: string
        default: demo-customer
      description: Customer identifier (defaults to demo-customer)

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: order-service
                  status:
                    type: string
                    example: ok

  /api/basket:
    get:
      tags: [Basket]
      summary: View current basket
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
      responses:
        "200":
          description: Current basket contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Basket"

  /api/basket/items:
    post:
      tags: [Basket]
      summary: Add item to basket
      description: Fetches cake details from catalog service and adds to basket.
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToBasketRequest"
      responses:
        "201":
          description: Item added to basket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Basket"
        "400":
          description: Invalid request (missing cakeId or invalid quantity)

  /api/basket/items/{cakeId}:
    patch:
      tags: [Basket]
      summary: Update item quantity
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
        - name: cakeId
          in: path
          required: true
          schema:
            type: string
          description: Cake ID of the item to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuantityRequest"
      responses:
        "200":
          description: Quantity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Basket"
        "404":
          description: Item not found in basket
    delete:
      tags: [Basket]
      summary: Remove item from basket
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
        - name: cakeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Basket"

  /api/orders/checkout:
    post:
      tags: [Orders]
      summary: Checkout basket and create order
      description: |
        Creates an order from the current basket, clears the basket,
        and publishes an `ORDER_COMPLETED` event to the Kafka `order-events` topic.
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Basket is empty or email missing

  /api/orders:
    get:
      tags: [Orders]
      summary: List orders for current customer
      parameters:
        - $ref: "#/components/parameters/XCustomerId"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found